+++ Sources/TuistKit/GraphMappers/ProjectMapperProvider.swift
@@ -44,6 +44,9 @@ final class ProjectMapperProvider: ProjectMapperProviding {
 
         // Project name and organization
         mappers.append(ProjectNameAndOrganizationMapper(config: config))
+        
+        // Development region
+        mappers.append(DevelopmentRegionMapper(config: config))
 
         // Signing
         mappers.append(SigningMapper())
